+++ lang/en/block_cmanager.php
@@ -104,6 +104,10 @@ $string['email_handover'] = 'Handover request';
 $string['email_noReply'] = 'Communications e-mail address';
 $string['email_noReplyInstructions'] = 'This utility sends email requests to administrators and end users as course requests are processed. Please enter the email address that these emails will appear to be sent from. It is recommended to use a no-reply email address. e.g. noreply@example.com';
 
+// Plugin settings.
+$string['norequestcontrol'] = 'Disable request control';
+$string['norequestcontrol_desc'] = 'Check this box to disable the option to disable the option to request control of an existing course.';
+
 //email subjects and contents
 $string['emailSubj_userApproved'] = "Moodle request approved!";
 $string['emailSubj_adminApproved'] = "Moodle user request approved!";